Move DiscreteSymmetry out of MFEMTopology header file
